+++ src/core/judgingthread.cpp
@@ -82,7 +82,7 @@ void JudgingThread::setInterpreterAsWatcher(bool use) { interpreterAsWatcher = u
 
 auto JudgingThread::getTimeUsed() const -> int { return timeUsed; }
 
-auto JudgingThread::getMemoryUsed() const -> int { return memoryUsed; }
+auto JudgingThread::getMemoryUsed() const -> qint64 { return memoryUsed; }
 
 auto JudgingThread::getScore() const -> int { return score; }
