+++ src/judgingdialog.h
@@ -43,7 +43,7 @@ class JudgingDialog : public QDialog {
 
   public slots:
 	void dialogAlert(const QString &);
-	void singleCaseFinished(QString, int, int, int, int, int, int, int);
+	void singleCaseFinished(QString, int, int, int, int, int, int, qint64);
 	void singleSubtaskDependenceFinished(int, int, int);
 	void taskJudgingStarted(const QString &);
 	void taskJudgedDisplay(const QString &, const QList<QList<int>> &, const int);
